+++ package.json
@@ -44,7 +44,6 @@
     "typescript": "^4.9.3"
   },
   "dependencies": {
-    "@octokit/rest": "^19.0.5",
     "async-retry": "^1.3.3",
     "markdown-table": "^2.0.0",
     "nock": "^13.2.9",
