+++ src/checker/MarkdownUtils.ts
