+++ src/index.ts
@@ -1,7 +1,7 @@
 import { ApplicationFunction, Probot } from 'probot'
 import { ApplicationFunctionOptions } from 'probot/lib/types'
 import { parseIdentifiers } from './utils'
-import { getCidChecker } from './dependency'
+import { getCidChecker } from './Dependency'
 
 const handler: ApplicationFunction = (app: Probot, options: ApplicationFunctionOptions): void => {
   app.log('Yay, the app was loaded!')
